chore(browser): create index file
